Add possibility to prefill the login input values (user / password)
Adds the interface to pass to the widget the values to be filled
at startup for login, password and (for completeness) confirmation
password.
This gives the possibility to prefill the form with saved accounts
or such, simplifying the login in such cases.
Note: saving and management of the accounts is not done in the plugin
but has to be done by the user (ie. using a SharedPreferences or so),
just the interface to manipulate the values is exposed here.